% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bestSubset.r
\name{invLog}
\alias{invLog}
\title{Inverse log likelihood function}
\usage{
invLog(fitness_metrics)
}
\arguments{
\item{fitness_metrics}{These metrics will be provided automatically by the fitnessMetric parameter in bestSubset()}
}
\value{
The function returns a vector with the inverse log of the given vector.
}
\description{
The inverse log function is for use as the "likelihoodFn" in bestSubset().  It will be greatest with smaller input values and is useful for minimizing model descriptive statistics such as error variance or prediction error. NOTE: minimize_fitness_metric in bestSubset() should be set to TRUE (the default) when using this likelihood function.
}
\examples{
See help for bestSubset()
}

